made sure that the solver for C++ default models converts all int inputs to ints

+++ rsnapsim/Solver.py
@@ -365,11 +365,11 @@ class Solver():
                     # PARAMETERS, BURNIN, TIME, AND KELONG MUST BE DOUBLES,
                     # everything else int32 or int.
                     mRNA_model.cmodel.run_ssa_cpp(pa, sa, ra, mRNA_model._rib_arr0, temp_state_arr0, temp_resource_arr0,
-                                               mRNA_model._rxn_mat.astype(np.int32), np.array(mRNA_model._constant_reactions + mRNA_model._ribosome_reactions),
+                                               mRNA_model._rxn_mat.astype(np.int32), np.array(mRNA_model._constant_reactions + mRNA_model._ribosome_reactions, dtype=np.int32),
                                                mRNA_model._kelong_mat,
                                                mRNA_model._probe_mat.astype(np.int32), np.array(pars),
                                                t.astype(np.float64), #convert t to a double
-                                               mRNA_model._rib_arr0.shape[0], mRNA_model._n_states, mRNA_model._n_resources, len(mRNA_model._constant_reactions),
+                                               int(mRNA_model._rib_arr0.shape[0]), int(mRNA_model._n_states), int(mRNA_model._n_resources), len(mRNA_model._constant_reactions),
                                                len(mRNA_model._ribosome_reactions),
                                                float(burnin), seed, )
